moved UnknownGrammarType to new exceptions package
