+++ src/edu/jhu/thrax/util/exceptions/UnknownGrammarTypeException.java
@@ -1,4 +1,4 @@
-package edu.jhu.thrax.util;
+package edu.jhu.thrax.util.exceptions;
 
 /**
  * This exception is thrown when the "grammar" option provided by the user is
